+++ lib/ui/favourites/favourites_screen.dart
@@ -1,8 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:lingo_jam/ui/favourites/favourites_view.dart';
-import 'package:lingo_jam/ui/style/style_constants.dart';
-import 'package:lingo_jam/ui/widgets/player.dart';
-import 'package:lingo_jam/ui/widgets/bottom_bar.dart';
+import 'package:lingo_jam/ui.dart';
 
 class FavouritesScreen extends StatelessWidget {
   const FavouritesScreen({Key? key}) : super(key: key);
