+++ lib/ui/stations/stations_screen.dart
@@ -1,8 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:lingo_jam/ui/stations/stations_view.dart';
-import 'package:lingo_jam/ui/style/style_constants.dart';
-import 'package:lingo_jam/ui/widgets/player.dart';
-import 'package:lingo_jam/ui/widgets/bottom_bar.dart';
+import 'package:lingo_jam/ui.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({Key? key}) : super(key: key);
