+++ lib/ui/test/test_screen.dart
@@ -1,9 +1,7 @@
 import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:lingo_jam/services/shared_preferences/blacklist_service.dart';
-import 'package:lingo_jam/ui/style/style_constants.dart';
-import 'package:lingo_jam/ui/widgets/bottom_bar.dart';
-import 'package:lingo_jam/ui/widgets/player.dart';
+import 'package:lingo_jam/ui.dart';
 
 class TestScreen extends StatelessWidget {
   const TestScreen({Key? key}) : super(key: key);
