+++ lib/ui/widgets/player.dart
@@ -2,8 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:lingo_jam/core/providers/app_state.dart';
 import 'package:lingo_jam/core/enums/playing_state.dart';
-import 'package:lingo_jam/ui/widgets/custom_card.dart';
-import 'package:lingo_jam/ui/widgets/widgets.dart';
+import 'package:lingo_jam/ui.dart';
 
 class PlayerBar extends StatelessWidget {
   const PlayerBar({Key? key}) : super(key: key);
