+++ lib/ui/widgets/widgets.dart
@@ -7,32 +7,46 @@ import 'package:lingo_jam/core/enums/playing_state.dart';
 import 'package:lingo_jam/core/providers/app_state.dart';
 import 'package:lingo_jam/core/providers/favourites_state.dart';
 import 'package:lingo_jam/model/station/station.dart';
-import 'package:lingo_jam/ui/widgets/tile.dart';
 
-class PlayPauseButton extends StatelessWidget {
-  const PlayPauseButton({Key? key}) : super(key: key);
+class CoverImage extends StatelessWidget {
+  const CoverImage(
+      {Key? key, required this.imageUrl, required this.placeholderImagePath})
+      : super(key: key);
 
-  final double iconSize = 40.0;
+  final String imageUrl;
+  final String placeholderImagePath;
 
   @override
   Widget build(BuildContext context) {
-    AppState state = context.watch<AppState>();
-    Widget _playPauseButton = Container();
+    return Container(
+      child: _backgroundImage(imageUrl, placeholderImagePath),
+    );
+  }
 
-    if (state.playingState == PlayingState.playing) {
-      _playPauseButton = IconButton(
-        iconSize: iconSize,
-        onPressed: state.pausePlaying,
-        icon: const Icon(Icons.pause_circle_filled),
-      );
-    } else if (state.playingState == PlayingState.paused) {
-      _playPauseButton = IconButton(
-        iconSize: iconSize,
-        onPressed: state.startPlaying,
-        icon: const Icon(Icons.play_circle),
-      );
+  final ImageProvider _placeholderImage =
+      const AssetImage("assets/images/vinyl-record-grey.png");
+
+  FadeInImage _backgroundImage(String url, String placeholderImagePath) {
+    if(url == "" || url.isEmpty) {
+      return FadeInImage(placeholder: _placeholderImage, image: AssetImage(placeholderImagePath));
     }
-    return _playPauseButton;
+
+    if (url.startsWith('http')) {
+      return FadeInImage(
+          placeholder: AssetImage(placeholderImagePath),
+          image: NetworkImage(url),
+          fit: BoxFit.contain,
+          imageErrorBuilder: (_, _a, _b) {
+            return Image.asset(placeholderImagePath);
+          });
+    }
+    return FadeInImage(
+        placeholder: AssetImage(placeholderImagePath),
+        image: AssetImage(url),
+        imageErrorBuilder: (_, _a, _b) {
+          debugPrint(placeholderImagePath + " url: " + url);
+          return Image.asset(placeholderImagePath);
+        });
   }
 }
 
@@ -63,43 +77,6 @@ class FavouriteButton extends StatelessWidget {
         icon,
         color: color,
       ),
-    );
-  }
-}
-
-class RemoveButton extends StatelessWidget {
-  const RemoveButton({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return TextButton(
-      child: const Text(
-        'Remove',
-        style: TextStyle(color: Colors.red),
-      ),
-      onPressed: context.watch<AppState>().removeAndBlacklistStream,
-    );
-  }
-}
-
-class PlayerButtonBar extends StatelessWidget {
-  const PlayerButtonBar({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    List<Widget> _children = [
-      const PlayPauseButton(),
-      const FavouriteButton(),
-    ];
-    if (context.watch<AppState>().playingState == PlayingState.errored) {
-      _children = [
-        const RemoveButton(),
-      ];
-    }
-
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.end,
-      children: _children,
     );
   }
 }
@@ -136,7 +113,7 @@ class SongNameAndPlayerButtonBar extends StatelessWidget {
             ),
           ),
         ),
-        const PlayerButtonBar(),
+        const _PlayerButtonBar(),
       ],
     );
   }
@@ -171,6 +148,72 @@ class StationNameAndImage extends StatelessWidget {
           ),
         )
       ],
+    );
+  }
+}
+
+
+// Private
+class _PlayerButtonBar extends StatelessWidget {
+  const _PlayerButtonBar({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    List<Widget> _children = [
+      const _PlayPauseButton(),
+      const FavouriteButton(),
+    ];
+    if (context.watch<AppState>().playingState == PlayingState.errored) {
+      _children = [
+        const _RemoveButton(),
+      ];
+    }
+
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.end,
+      children: _children,
+    );
+  }
+}
+
+class _PlayPauseButton extends StatelessWidget {
+  const _PlayPauseButton({Key? key}) : super(key: key);
+
+  final double iconSize = 40.0;
+
+  @override
+  Widget build(BuildContext context) {
+    AppState state = context.watch<AppState>();
+    Widget _playPauseButton = Container();
+
+    if (state.playingState == PlayingState.playing) {
+      _playPauseButton = IconButton(
+        iconSize: iconSize,
+        onPressed: state.pausePlaying,
+        icon: const Icon(Icons.pause_circle_filled),
+      );
+    } else if (state.playingState == PlayingState.paused) {
+      _playPauseButton = IconButton(
+        iconSize: iconSize,
+        onPressed: state.startPlaying,
+        icon: const Icon(Icons.play_circle),
+      );
+    }
+    return _playPauseButton;
+  }
+}
+
+class _RemoveButton extends StatelessWidget {
+  const _RemoveButton({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return TextButton(
+      child: const Text(
+        'Remove',
+        style: TextStyle(color: Colors.red),
+      ),
+      onPressed: context.watch<AppState>().removeAndBlacklistStream,
     );
   }
 }
